requirements.txt
matplotlib==3.1.2
numpy==1.17.3
Keras==2.2.5
streamlit==0.52.2
Pillow==6.2.1
tensorflow==1.14.0
